Moved tests from UserDocumentControllerTest (deleted) to UserControllerTest
